Move Worley to Core
