+++ src/core.rs
@@ -1,3 +1,4 @@
 pub mod perlin;
-pub mod simplex;
 pub mod perlin_surflet;
+pub mod simplex;
+pub mod worley;
